refactor(database): split raw flash backend internals
